fix(content API): Refactor to use global + local flags in comination
Global flags, vars, stats for clear shared states. Separate files to communicate intention.

Local flags and vars are still encouraged to use whenever possible.
