+++ scripts/scenarios/default/stats.ts
